Alignment fix; found the hard way on sparc64 by landry@ (and thanks too for fast access to a box!)
ok deraadt@, matthieu@

+++ app/cwm/xutil.c
@@ -15,7 +15,7 @@
  * ACTION OF CONTRACT, NEGLIGENCE OR OTHER TORTIOUS ACTION, ARISING OUT OF
  * OR IN CONNECTION WITH THE USE OR PERFORMANCE OF THIS SOFTWARE.
  *
- * $OpenBSD: xutil.c,v 1.84 2014/02/07 21:59:56 okan Exp $
+ * $OpenBSD: xutil.c,v 1.85 2014/02/27 00:52:57 okan Exp $
  */
 
 #include <sys/param.h>
@@ -291,7 +291,7 @@ xu_ewmh_net_desktop_names(struct screen_ctx *sc, char *data, int n)
 void
 xu_ewmh_net_wm_desktop(struct client_ctx *cc)
 {
-	int	 no = cc->group->shortcut;
+	long	 no = cc->group->shortcut;
 
 	XChangeProperty(X_Dpy, cc->win, ewmh[_NET_WM_DESKTOP],
 	    XA_CARDINAL, 32, PropModeReplace, (unsigned char *)&no, 1);
